Enable low latency tcp by default
In the general case I have been going for low latency
over throughput. I have to note that to do more realistic
tests, turning off some of the low latency features such
as this one would be a good idea for some tests.

However, 'low latency by default' is to be preferred
for now.