Put new sfq flows to head.
Major conflict between science and engineering here.
Without fiddling with this patch at least a little, coming up with ways
to make it better.

This patch made sfq perform as well as qfq.

And as qfq is busted, and the circumstances under which this bug
occured I was unable to duplicate in cerowrt (I have no doubt it
exists), I'd like to duplicate it and see if fixing htb helps.
Later on.

I will revert this patch if I can create pathological circumstances
with the cerowrt code.