ipv6: Various ipv6 and ipv4 functions threw alignment traps
This is a comprehensive (and somewhat invasive) set of fixes
for the problem. We went from 100k traps a second on ipv6
to .7/sec, and performance more than doubled.

Also, crashes in tunneling stopped.

ipv4: From 33k traps/sec on ipv4 to 8k/sec, and performance
improved by about 30Mbit/sec.

The only traps left are in the aqm code, which ironically
is what we need to exercise most. So fixing those is
next.