Enable TCP timestamps. Enable sack/dsack. Enable ECN if offered
A year of testing in the cerowrt project shows not using timestamps
to be a very bad idea in nearly any TCP at speeds above a few Mbit.

As FQ_CODEL will actually assert ECT(3) on ECN streams, it is useful
to have ECN be negotiated when offered. In this day and age it
could also be hoped that ECN could merely be one by default, but...

Lastly sack/dsack help on recovery from larger amounts of packet
loss.