qos-scripts reduce fq_codel limit significantly
The default fq_codel limit is 10k packets. For systems using
openwrt qos, the potential a udp flood, in particular, can
rapidly run a machine out of memory. 1k packets is enough
for most normal usages.

This patch also (experimentally) alters the quantum to 505 bytes
in the hope that this will result in better mixing.