Memory use improvement hacks
These cut memory use in codel and fq_codel under load enormously.

I am unsure if there is actually enough headroom reallocated by default
in the skb, some code in the ath9k driver appears to be wrong about
it's handling of headroom. Packets lacking enough seemed to be dropped
although I haven't seen this in the real world.

It would also be sane to share max ath9k headroom
values with the ar71xx ethernet driver so that no reallocations would ever
be necessary between ethernet and wifi.

(this might be a win in the general case)

There is another patch by eric dumazet that cuts the std ath9k allocation
in half, but I was dubious about it, and haven't tried it.